address traversal edge case
This set of changes ensures the correct traversal of a multiring
wherein a data node is separated from the next data node or the root
node by one or more rings.

The findLast method on the various types of node is reimplemented more
efficiently on the MultiRing type itself.

The GateNode.exit method is no longer required and is dropped.

BREAKING CHANGE: remove findLast method on Node, GateNode, DataNode
BREAKING CHANGE: remove exit method on GateNode