add memory management example
Consumers of this module will have differing requirements and therefore
differing opinions about the design of a higher-level API over the core
ADT. Differences in requirements will affect, for example, the choice
and implementation of algorithms for sorting, how memory is managed and
how undefined behavior is handled.

So, instead of extending the core ADT, we add an example that shows how
the core ADT, which contains only essential complexity, can be embedded
into a type containing a mixture of accidental and essential complexity.

Relates to #5