AutoMultiRing: add filter, fold and reduce methods
These methods apply to the entire multiring, not to individual rings, to
illustrate the ease with which we can leverage the core ADT's API.

The filter method allows us to omit rings of arbitrary size
conditionally when traversing the multiring. It highlights how the
structure of the multiring can decrease the time complexity of traversal
without increasing the space complexity.

Relates to #5